add error handling on delete and email exist error

+++ src/main/java/com/example/demo/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.example.demo.student;
 
+import com.example.demo.exception.BadRequestException;
+import com.example.demo.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -9,17 +11,31 @@ import java.util.List;
 @AllArgsConstructor
 public class StudentService {
 
+
+    public static long count=0;
     private final StudentRepository studentRepository;
+
 
     public List<Student> getAllStudents(){
         return studentRepository.findAll();
     }
 
+    public long countEmail(String email){
+        count = studentRepository.count(email);
+        return count;
+    }
     public void addStudent(Student student) {
-        studentRepository.save(student);
+        if(countEmail(student.getEmail())>0) {
+            throw new BadRequestException("Email with Student already exist");
+        } else {
+            studentRepository.save(student);
+        }
     }
 
     public void deleteStudent(long studentId){
+        if(!studentRepository.existsById(studentId)){
+            throw new StudentNotFoundException("Student with Id: "+studentId+" does not exist");
+        }
         studentRepository.deleteById(studentId);
     }
 }
